app/requirements.txt
psycopg2-binary
pandas
openpyxl
pyspark==4.0.0
pyarrow>=11.0.0
grpcio>=1.67.0
grpcio-status>=1.67.0
googleapis-common-protos>=1.65.0


